+++ NEWS.md
@@ -1,9 +1,11 @@
 # ibis.iSDM 0.1.1 (current dev branch)
 
 #### New features
+* Added default `engine_glm()` for dependency-free inference and projection.
 * Harmonized controls settings and added option to contrain extrapolation `add_control_extrapolation()`
 
 #### Minor improvements and bug fixes
+* Switch to `engine_glm()` in many of the unittests for better coverage.
 * Several bug fixes and improvements in `thin_observations`
 * `global`, `probs`, and `centers` argument for better control of `thin_observations`
 * Harmonization of parameters for `spartial()` and addressing #80